Don't create fifo if already present
When using S6_BEHAVIOUR_IF_STAGE2_FAILS=2, a container will not start if any of
the stage2 scripts fail.

This is caused by the fact that `00-initialize-log-fds` script will fail if one
of the fifos has already been created. For example, this is the case when
restarting the container.

This commits checks if the fifo exists before trying to create it (using
`s6-test -p`).